Fix comment
uri.host is used for the "Host: ..." field in the prestarted HEAD request, which the user might purposely wish to be set to a a load balancer address (the library also specifies it should match the server name of the segment where the app is that you want to prestart). But you do not want to connect to a loadbalancer address for prestarting.